feat(worklist): a row leads to the record it is about
Every lane on the Worklist already knew which record its row concerned - the
deal that went quiet, the person a promise was made to, the lead a task was
raised for - and drew it as text. A rep who read 'Salesforce - Einfuehrung, no
contact for 83 days' then had to go and find Salesforce by hand, so the queue
named work it would not let anyone start.

The task lane now carries the record it is filed under. An activity may be
filed under several, so primaryLink picks one by a stated priority (lead, deal,
project, person, organization) rather than by whatever order the store
returned: the same task must not lead to the company on one read and the lead
on the next. The links it reads are already row-scope filtered by the
activities read, so a record the reader may not see never becomes a
destination.

The quiet-deal and promise cards now send the "open" action. Both name a record
with a page, and warning someone about a debt while withholding the way to pay
it is the defect this fixes. The meetings card still sends no verb, for a
narrower reason than before: its subject is the ACTIVITY behind the
appointment, a timeline entry with no page of its own.

Two gates hold the honesty. The existing one still refuses an action the
surface cannot render; the new one refuses "open" on a card whose subject is
not a record that can be opened, which is what would otherwise let a lane
advertise navigation to nowhere.
