+++ backend/internal/compose/attention/lanewiring_test.go
@@ -5,6 +5,7 @@ package attention
 
 import (
 	"context"
+	"slices"
 	"testing"
 
 	crmcontracts "github.com/margince/margince/backend/internal/contracts"
@@ -49,17 +50,21 @@ func TestEachOptionalLaneFillsItsOwnField(t *testing.T) {
 
 // No lane advertises an action the surface cannot perform.
 //
-// AttentionRow renders controls for `complete` and `snooze` only. A card
-// offering anything else is a promise to a client that nothing keeps — and a
-// generated client is entitled to draw a control for whatever the server says
-// it offers. The three optional lanes therefore send no action at all until the
-// navigation they would need actually exists.
+// A card offering an action nothing renders is a promise to a client that
+// nothing keeps — a generated client is entitled to draw a control for whatever
+// the server says it offers. AttentionRow renders `complete` and `snooze` as
+// buttons, and `open` as the row's own link to the record named in `subject`.
+//
+// `open` is performable only where the subject is a record with a page. The
+// meetings lane's subject is the ACTIVITY behind the appointment, which is a
+// timeline entry and not a destination, which is why that lane still sends no
+// action at all.
 //
 // Derived from the lanes rather than listed, so a fourth joins the check by
 // existing.
 func TestNoOptionalLaneOffersAnActionTheSurfaceCannotPerform(t *testing.T) {
 	performable := map[crmcontracts.AttentionItemActions]bool{
-		"complete": true, "snooze": true,
+		"complete": true, "snooze": true, "open": true,
 	}
 	svc := NewService(
 		stubApprovals{}, stubDuplicates{}, &stubTasks{}, stubReceipts{}, stubBriefing{},
@@ -86,6 +91,56 @@ func TestNoOptionalLaneOffersAnActionTheSurfaceCannotPerform(t *testing.T) {
 					t.Errorf("%s offers %q, which no control on this surface performs",
 						name, action)
 				}
+			}
+		}
+	}
+}
+
+// An item offering `open` names a record the reader can open.
+//
+// `open` means "go to the thing this card is about", so it is only honest when
+// the card carries the thing: a subject, of a type that is a record with a page
+// rather than an entry on a timeline. Without this the previous gate would pass
+// for a lane that advertised navigation to nowhere, which is the exact defect
+// that kept every one of these lanes inert.
+func TestOpenIsOfferedOnlyWithARecordToOpen(t *testing.T) {
+	// `activity` is deliberately absent: it is a timeline entry, and no screen
+	// answers to it.
+	openable := map[crmcontracts.AttentionSubjectType]bool{
+		"organization": true, "person": true, "deal": true,
+		"lead": true, "project": true,
+	}
+	svc := NewService(
+		stubApprovals{}, stubDuplicates{}, &stubTasks{}, stubReceipts{}, stubBriefing{},
+		&stubCommitments{rows: []Commitment{promise("a promise", readInstant)}},
+		stubAtRisk{rows: []RiskyDeal{{Name: "a deal", QuietDays: 20}}},
+		&stubDecay{rows: []QuietRelationship{{Name: "a contact", QuietDays: 63, LastAt: readInstant}}},
+		&stubMeetings{rows: []Meeting{{Subject: "a meeting", StartsAt: readInstant}}},
+		fixedClock,
+	)
+	out, err := svc.Assemble(context.Background())
+	if err != nil {
+		t.Fatalf("assembling: %v", err)
+	}
+	lanes := map[string]*[]crmcontracts.AttentionItem{
+		"meetings": out.Meetings, "at_risk": out.AtRisk,
+		"commitments": out.Commitments, "planned": &out.Planned,
+	}
+	for name, lane := range lanes {
+		if lane == nil {
+			t.Fatalf("%s is absent, so this gate checked nothing", name)
+		}
+		for _, item := range *lane {
+			if !slices.Contains(item.Actions, "open") {
+				continue
+			}
+			if item.Subject == nil {
+				t.Errorf("%s offers open with no subject to open", name)
+				continue
+			}
+			if !openable[item.Subject.Type] {
+				t.Errorf("%s offers open on a %q, which has no page",
+					name, item.Subject.Type)
 			}
 		}
 	}
