+++ frontend/src/screens/today.css
@@ -53,8 +53,29 @@
   gap: var(--space-3);
 }
 
+/* A verb-less row is one press target filling the row, so the anchor has to be
+ * the thing that fills it — inheriting the row's layout rather than sitting as
+ * an inline word inside it. */
+.today-row-whole {
+  width: 100%;
+  color: inherit;
+  text-decoration: none;
+}
+
 .today-row-text {
   min-width: 0;
+}
+
+/* On a row that already carries verbs the NAME is the link, and it says so the
+ * way every other in-text link in the product does. */
+.today-row-link {
+  color: inherit;
+  text-decoration: underline;
+  text-underline-offset: 2px;
+}
+
+.today-row-link:hover {
+  color: var(--accent);
 }
 
 .today-row-title {
